feat: camera access denied message
