+++ packages/camera/camera_web/lib/camera_web.dart
@@ -1,4 +1,3 @@
 library camera_web;
 
-export 'src/camera_exception.dart';
 export 'src/camera_web.dart' show CameraPlugin;
